+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## unreleased
 
+* CSI driver can accidentally format existing volume
+  [[GH-469]](https://github.com/digitalocean/csi-digitalocean/pull/478)
+
 ## v4.5.0 - 2023.01.11
 * Update CSI driver for Kubernetes 1.26
   [[GH-469]](https://github.com/digitalocean/csi-digitalocean/pull/469)